add shortcuts script, tweak nvim configs, format scripts, change sxhkd key bindings
add shortcuts script to modify the display name of the directory in
shell and generate shortcuts based on the bm-dirs and bm-files

change how gen-xcolor script get the xresources color and add the
colorscheme name of that colors.

tweak x11 xresources formatting

add shell aliases and add logic to some, so it does not break when certain
program is not installed.

add shell env var for xdg state home

change some sxhkd key bindings and tweak the config file formatting